+++ frontend/src/pages/Projects/Project/ProjectDetails/BudgetSection/formConfig.ts
@@ -221,9 +221,8 @@ export const formConfig = (query: UseQueryResult<AxiosResponse, unknown>) => {
   };
 
   const rowsToLock = [query?.data?.data?.data?.id];
-  const queryKey = `/projects/budget/${query?.data?.data?.data?.id}`;
   const postUrl = `/projects/budget`;
   const updateUrl = `/projects/budget/${query?.data?.data?.data?.id}`;
 
-  return { readFields, editFields, initialValues, rowsToLock, queryKey, postUrl, updateUrl };
+  return { readFields, editFields, initialValues, rowsToLock, postUrl, updateUrl };
 };