+++ frontend/src/pages/Projects/Project/ProjectDetails/ClientCodingSection/formConfig.ts
@@ -115,9 +115,8 @@ export const formConfig = (query: UseQueryResult<AxiosResponse, unknown>) => {
   };
 
   const rowsToLock = [query?.data?.data?.data?.id];
-  const queryKey = `/projects/client-coding/${query?.data?.data?.data?.id}`;
   const postUrl = `/projects/${query?.data?.data?.data?.project_id}/client-coding`;
   const updateUrl = `/projects/client-coding/${query?.data?.data?.data?.id}`;
 
-  return { readFields, editFields, initialValues, rowsToLock, queryKey, postUrl, updateUrl };
+  return { readFields, editFields, initialValues, rowsToLock, postUrl, updateUrl };
 };