+++ frontend/src/pages/Projects/Project/ProjectDetails/Deliverables/formConfig.ts
@@ -168,9 +168,8 @@ export const formConfig = (query: UseQueryResult<AxiosResponse, unknown>) => {
   };
 
   const rowsToLock = [query?.data?.data?.data?.id];
-  const queryKey = `/projects/deliverables/${query?.data?.data?.data?.id}`;
   const postUrl = `projects/deliverables`;
   const updateUrl = `projects/deliverables/${query?.data?.data?.data?.id}`;
 
-  return { readFields, editFields, initialValues, rowsToLock, queryKey, postUrl, updateUrl };
+  return { readFields, editFields, initialValues, rowsToLock, postUrl, updateUrl };
 };