+++ frontend/src/pages/Projects/Project/Status/formConfig.ts
@@ -146,9 +146,8 @@ export const formConfig = (query: UseQueryResult<AxiosResponse, unknown>) => {
   };
 
   const rowsToLock = [query?.data?.data?.data?.id];
-  const queryKey = `/projects/status/${query?.data?.data?.data?.id}`;
   const postUrl = `/projects/status`;
   const updateUrl = `/projects/status/${query?.data?.data?.data?.id}`;
 
-  return { readFields, editFields, initialValues, rowsToLock, queryKey, postUrl, updateUrl };
+  return { readFields, editFields, initialValues, rowsToLock, postUrl, updateUrl };
 };