+++ CHANGELOG.md
@@ -5,7 +5,9 @@
 - Frontend
 
   - Added db_lock to project close out section.
-
+  - Fixed render bugs and styling for table component
+  - Updated how react query for the frontend works with form and table data
+  
 **May 23, 2023**
 
 - Backend